remove PIR model as it's already in there, impute PIR based on non-res model
