+++ jobs/archive_filled_posts_estimates.py
@@ -30,7 +30,7 @@ MONTHLY_ARCHIVE_COLUMNS = [
     IndCQC.estimate_filled_posts,
     IndCQC.estimate_filled_posts_source,
     IndCQC.imputed_gac_service_types,
-    IndCQC.imputed_non_res_pir_people_directly_employed,
+    IndCQC.imputed_pir_filled_posts_model,
     IndCQC.imputed_posts_care_home_model,
     IndCQC.imputed_posts_non_res_combined_model,
     IndCQC.imputed_filled_post_model,
@@ -44,6 +44,7 @@ MONTHLY_ARCHIVE_COLUMNS = [
     IndCQC.number_of_beds,
     IndCQC.organisation_id,
     IndCQC.pir_people_directly_employed_dedup,
+    IndCQC.pir_filled_posts_model,
     IndCQC.primary_service_type,
     IndCQC.provider_id,
     IndCQC.provider_name,
