+++ jobs/estimate_ind_cqc_filled_posts.py
@@ -20,9 +20,6 @@ from utils.estimate_filled_posts.models.non_res_without_dormancy import (
 from utils.estimate_filled_posts.models.non_res_with_and_without_dormancy_combined import (
     combine_non_res_with_and_without_dormancy_models,
 )
-from utils.estimate_filled_posts.models.non_res_pir_linear_regression import (
-    model_non_res_pir_linear_regression,
-)
 from utils.ind_cqc_filled_posts_utils.utils import merge_columns_in_order
 
 ind_cqc_columns = [
@@ -43,6 +40,7 @@ ind_cqc_columns = [
     IndCQC.registered_manager_names,
     IndCQC.cqc_pir_import_date,
     IndCQC.pir_people_directly_employed_dedup,
+    IndCQC.pir_filled_posts_model,
     IndCQC.ascwds_workplace_import_date,
     IndCQC.establishment_id,
     IndCQC.organisation_id,
@@ -61,7 +59,6 @@ ind_cqc_columns = [
     IndCQC.current_rural_urban_indicator_2011,
     IndCQC.posts_rolling_average_model,
     IndCQC.imputed_filled_post_model,
-    IndCQC.imputed_non_res_pir_people_directly_employed,
     IndCQC.imputed_filled_posts_per_bed_ratio_model,
     IndCQC.unix_time,
     Keys.year,
@@ -81,8 +78,6 @@ def main(
     non_res_with_dormancy_model_source: str,
     non_res_without_dormancy_features_source: str,
     non_res_without_dormancy_model_source: str,
-    non_res_pir_linear_regression_features_source: str,
-    non_res_pir_linear_regression_model_source: str,
     estimated_ind_cqc_destination: str,
     ml_model_metrics_destination: str,
 ) -> DataFrame:
@@ -101,9 +96,6 @@ def main(
     )
     non_res_without_dormancy_features_df = utils.read_from_parquet(
         non_res_without_dormancy_features_source
-    )
-    non_res_pir_linear_regression_features_df = utils.read_from_parquet(
-        non_res_pir_linear_regression_features_source
     )
 
     estimate_filled_posts_df = model_care_homes(
@@ -130,13 +122,6 @@ def main(
         estimate_filled_posts_df
     )
 
-    estimate_filled_posts_df = model_non_res_pir_linear_regression(
-        estimate_filled_posts_df,
-        non_res_pir_linear_regression_features_df,
-        non_res_pir_linear_regression_model_source,
-        ml_model_metrics_destination,
-    )
-
     estimate_filled_posts_df = model_imputation_with_extrapolation_and_interpolation(
         estimate_filled_posts_df,
         IndCQC.ascwds_pir_merged,
@@ -153,7 +138,13 @@ def main(
         care_home=False,
     )
 
-    # TODO: add imputation for other non res models
+    estimate_filled_posts_df = model_imputation_with_extrapolation_and_interpolation(
+        estimate_filled_posts_df,
+        IndCQC.pir_filled_posts_model,
+        IndCQC.non_res_combined_model,
+        IndCQC.imputed_pir_filled_posts_model,
+        care_home=False,
+    )
 
     estimate_filled_posts_df = merge_columns_in_order(
         estimate_filled_posts_df,
@@ -162,7 +153,7 @@ def main(
             IndCQC.imputed_posts_care_home_model,
             IndCQC.care_home_model,
             IndCQC.imputed_posts_non_res_combined_model,
-            IndCQC.non_res_pir_linear_regression_model,
+            IndCQC.imputed_pir_filled_posts_model,
             IndCQC.non_res_combined_model,
             IndCQC.posts_rolling_average_model,
         ],
@@ -194,8 +185,6 @@ if __name__ == "__main__":
         non_res_with_dormancy_model_source,
         non_res_without_dormancy_features_source,
         non_res_without_dormancy_model_source,
-        non_res_pir_linear_regression_features_source,
-        non_res_pir_linear_regression_model_source,
         estimated_ind_cqc_destination,
         ml_model_metrics_destination,
     ) = utils.collect_arguments(
@@ -228,14 +217,6 @@ if __name__ == "__main__":
             "Source s3 directory for the non res without dormancy ML model",
         ),
         (
-            "--non_res_pir_linear_regression_features_source",
-            "Source s3 directory for non res pir linear regression features dataset",
-        ),
-        (
-            "--non_res_pir_linear_regression_model_source",
-            "Source s3 directory for the non res pir linear regression model",
-        ),
-        (
             "--estimated_ind_cqc_destination",
             "Destination s3 directory for outputting estimates for filled posts",
         ),
@@ -253,8 +234,6 @@ if __name__ == "__main__":
         non_res_with_dormancy_model_source,
         non_res_without_dormancy_features_source,
         non_res_without_dormancy_model_source,
-        non_res_pir_linear_regression_features_source,
-        non_res_pir_linear_regression_model_source,
         estimated_ind_cqc_destination,
         ml_model_metrics_destination,
     )
