+++ projects/_03_independent_cqc/unittest_data/ind_cqc_test_file_schemas.py
@@ -130,6 +130,45 @@ class ModelAndMergePirData:
 
 
 @dataclass
+class ValidateImputedIndCqcAscwdsAndPir:
+    cleaned_ind_cqc_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), True),
+            StructField(IndCQC.cqc_location_import_date, DateType(), True),
+        ]
+    )
+    imputed_ind_cqc_ascwds_and_pir_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), True),
+            StructField(IndCQC.cqc_location_import_date, DateType(), True),
+            StructField(IndCQC.ascwds_workplace_import_date, DateType(), True),
+            StructField(IndCQC.cqc_pir_import_date, DateType(), True),
+            StructField(IndCQC.care_home, StringType(), True),
+            StructField(IndCQC.provider_id, StringType(), True),
+            StructField(IndCQC.cqc_sector, StringType(), True),
+            StructField(IndCQC.registration_status, StringType(), True),
+            StructField(IndCQC.imputed_registration_date, DateType(), True),
+            StructField(IndCQC.dormancy, StringType(), True),
+            StructField(IndCQC.number_of_beds, IntegerType(), True),
+            StructField(IndCQC.primary_service_type, StringType(), True),
+            StructField(IndCQC.contemporary_ons_import_date, DateType(), True),
+            StructField(IndCQC.contemporary_cssr, StringType(), True),
+            StructField(IndCQC.contemporary_region, StringType(), True),
+            StructField(IndCQC.current_ons_import_date, DateType(), True),
+            StructField(IndCQC.current_cssr, StringType(), True),
+            StructField(IndCQC.current_region, StringType(), True),
+            StructField(IndCQC.current_rural_urban_indicator_2011, StringType(), True),
+            StructField(IndCQC.total_staff_bounded, IntegerType(), True),
+            StructField(IndCQC.worker_records_bounded, IntegerType(), True),
+            StructField(IndCQC.ascwds_filled_posts_source, StringType(), True),
+            StructField(IndCQC.ascwds_filled_posts_dedup_clean, FloatType(), True),
+            StructField(IndCQC.pir_people_directly_employed_dedup, IntegerType(), True),
+        ]
+    )
+    calculate_expected_size_schema = cleaned_ind_cqc_schema
+
+
+@dataclass
 class TrainLinearRegressionModelSchema:
     feature_schema = StructType(
         [
