+++ terraform/pipeline/glue.tf
@@ -222,21 +222,6 @@ module "prepare_features_non_res_ascwds_ind_cqc_job" {
   }
 }
 
-
-module "prepare_features_non_res_pir_ind_cqc_job" {
-  source          = "../modules/glue-job"
-  script_dir      = "jobs"
-  script_name     = "prepare_features_non_res_pir_ind_cqc.py"
-  glue_role       = aws_iam_role.sfc_glue_service_iam_role
-  resource_bucket = module.pipeline_resources
-  datasets_bucket = module.datasets_bucket
-
-  job_parameters = {
-    "--ind_cqc_cleaned_data_source"              = "${module.datasets_bucket.bucket_uri}/domain=ind_cqc_filled_posts/dataset=ind_cqc_imputed_ascwds_and_pir/"
-    "--non_res_pir_ind_cqc_features_destination" = "${module.datasets_bucket.bucket_uri}/domain=ind_cqc_filled_posts/dataset=ind_cqc_features_non_res_pir/"
-  }
-}
-
 module "clean_ind_cqc_filled_posts_job" {
   source          = "../modules/glue-job"
   script_dir      = "jobs"
@@ -629,7 +614,7 @@ module "validate_cleaned_ind_cqc_data_job" {
 
 module "validate_imputed_ind_cqc_ascwds_and_pir_data_job" {
   source          = "../modules/glue-job"
-  script_dir      = "jobs"
+  script_dir      = "projects/_03_independent_cqc/_03_impute/jobs"
   script_name     = "validate_imputed_ind_cqc_ascwds_and_pir_data.py"
   glue_role       = aws_iam_role.sfc_glue_service_iam_role
   resource_bucket = module.pipeline_resources
@@ -689,23 +674,6 @@ module "validate_features_non_res_ascwds_without_dormancy_ind_cqc_data_job" {
     "--cleaned_ind_cqc_source"                                  = "${module.datasets_bucket.bucket_uri}/domain=ind_cqc_filled_posts/dataset=ind_cqc_cleaned_data/"
     "--non_res_ascwds_without_dormancy_ind_cqc_features_source" = "${module.datasets_bucket.bucket_uri}/domain=ind_cqc_filled_posts/dataset=ind_cqc_features_non_res_ascwds_without_dormancy/"
     "--report_destination"                                      = "${module.datasets_bucket.bucket_uri}/domain=data_validation_reports/dataset=data_quality_report_ind_cqc_features_non_res_ascwds_without_dormancy/"
-  }
-}
-
-
-module "validate_features_non_res_pir_ind_cqc_data_job" {
-  source          = "../modules/glue-job"
-  script_dir      = "jobs"
-  script_name     = "validate_features_non_res_pir_ind_cqc_data.py"
-  glue_role       = aws_iam_role.sfc_glue_service_iam_role
-  resource_bucket = module.pipeline_resources
-  datasets_bucket = module.datasets_bucket
-  glue_version    = "4.0"
-
-  job_parameters = {
-    "--cleaned_ind_cqc_source"              = "${module.datasets_bucket.bucket_uri}/domain=ind_cqc_filled_posts/dataset=ind_cqc_imputed_ascwds_and_pir/"
-    "--non_res_pir_ind_cqc_features_source" = "${module.datasets_bucket.bucket_uri}/domain=ind_cqc_filled_posts/dataset=ind_cqc_features_non_res_pir/"
-    "--report_destination"                  = "${module.datasets_bucket.bucket_uri}/domain=data_validation_reports/dataset=data_quality_report_non_res_pir_ind_cqc_features/"
   }
 }
 
@@ -849,7 +817,7 @@ module "prepare_features_care_home_ind_cqc_job" {
 
 module "impute_ind_cqc_ascwds_and_pir_job" {
   source            = "../modules/glue-job"
-  script_dir        = "jobs"
+  script_dir        = "projects/_03_independent_cqc/_03_impute/jobs"
   script_name       = "impute_ind_cqc_ascwds_and_pir.py"
   glue_role         = aws_iam_role.sfc_glue_service_iam_role
   worker_type       = "G.1X"
@@ -882,8 +850,6 @@ module "estimate_ind_cqc_filled_posts_job" {
     "--non_res_with_dormancy_model_source"            = "${module.pipeline_resources.bucket_uri}/models/non_residential_with_dormancy_prediction/4.0.0/"
     "--non_res_without_dormancy_features_source"      = "${module.datasets_bucket.bucket_uri}/domain=ind_cqc_filled_posts/dataset=ind_cqc_features_non_res_ascwds_without_dormancy/"
     "--non_res_without_dormancy_model_source"         = "${module.pipeline_resources.bucket_uri}/models/non_residential_without_dormancy_prediction/4.0.0/"
-    "--non_res_pir_linear_regression_features_source" = "${module.datasets_bucket.bucket_uri}/domain=ind_cqc_filled_posts/dataset=ind_cqc_features_non_res_pir/"
-    "--non_res_pir_linear_regression_model_source"    = "${module.pipeline_resources.bucket_uri}/models/non_res_pir_linear_regression_prediction/2.0.1/"
     "--estimated_ind_cqc_destination"                 = "${module.datasets_bucket.bucket_uri}/domain=ind_cqc_filled_posts/dataset=ind_cqc_estimated_filled_posts/"
     "--ml_model_metrics_destination"                  = "${module.datasets_bucket.bucket_uri}/domain=ind_cqc_filled_posts/dataset=ind_cqc_ml_model_metrics/"
   }
