+++ terraform/pipeline/step-function.tf
@@ -24,8 +24,6 @@ resource "aws_sfn_state_machine" "ind_cqc_filled_post_estimates_pipeline_state_m
     prepare_features_non_res_ascwds_ind_cqc_job_name                        = module.prepare_features_non_res_ascwds_ind_cqc_job.job_name
     validate_features_non_res_ascwds_with_dormancy_ind_cqc_data_job_name    = module.validate_features_non_res_ascwds_with_dormancy_ind_cqc_data_job.job_name
     validate_features_non_res_ascwds_without_dormancy_ind_cqc_data_job_name = module.validate_features_non_res_ascwds_without_dormancy_ind_cqc_data_job.job_name
-    prepare_features_non_res_pir_ind_cqc_job_name                           = module.prepare_features_non_res_pir_ind_cqc_job.job_name
-    validate_features_non_res_pir_ind_cqc_data_job_name                     = module.validate_features_non_res_pir_ind_cqc_data_job.job_name
     estimate_ind_cqc_filled_posts_job_name                                  = module.estimate_ind_cqc_filled_posts_job.job_name
     validate_estimated_ind_cqc_filled_posts_data_job_name                   = module.validate_estimated_ind_cqc_filled_posts_data_job.job_name
     estimate_ind_cqc_filled_posts_by_job_role_job_name                      = module.estimate_ind_cqc_filled_posts_by_job_role_job.job_name
