+++ tests/test_file_schemas.py
@@ -2345,26 +2345,6 @@ class CareHomeFeaturesSchema:
 
 
 @dataclass
-class NonResPirFeaturesSchema:
-    features_schema = StructType(
-        [
-            StructField(IndCQC.location_id, StringType(), True),
-            StructField(IndCQC.cqc_location_import_date, DateType(), True),
-            StructField(IndCQC.care_home, StringType(), True),
-            StructField(IndCQC.ascwds_pir_merged, FloatType(), True),
-            StructField(IndCQC.pir_people_directly_employed_dedup, IntegerType(), True),
-            StructField(
-                IndCQC.imputed_non_res_pir_people_directly_employed, FloatType(), True
-            ),
-            StructField(Keys.year, StringType(), True),
-            StructField(Keys.month, StringType(), True),
-            StructField(Keys.day, StringType(), True),
-            StructField(Keys.import_date, StringType(), True),
-        ]
-    )
-
-
-@dataclass
 class EstimateIndCQCFilledPostsSchemas:
     cleaned_ind_cqc_schema = StructType(
         [
@@ -3193,37 +3173,6 @@ class ModelNonResWithAndWithoutDormancyCombinedSchemas:
         [
             *combine_model_predictions_schema,
             StructField(IndCQC.prediction, FloatType(), True),
-        ]
-    )
-
-
-@dataclass
-class ModelNonResPirLinearRegressionSchemas:
-    non_res_pir_cleaned_ind_cqc_schema = StructType(
-        [
-            StructField(IndCQC.location_id, StringType(), True),
-            StructField(IndCQC.cqc_location_import_date, DateType(), True),
-            StructField(IndCQC.care_home, StringType(), True),
-            StructField(
-                IndCQC.imputed_non_res_pir_people_directly_employed, FloatType(), True
-            ),
-        ]
-    )
-    non_res_pir_features_schema = StructType(
-        [
-            StructField(IndCQC.location_id, StringType(), True),
-            StructField(IndCQC.cqc_location_import_date, DateType(), True),
-            StructField(IndCQC.care_home, StringType(), True),
-            StructField(
-                IndCQC.imputed_non_res_pir_people_directly_employed, FloatType(), True
-            ),
-            StructField(IndCQC.features, VectorUDT(), True),
-        ]
-    )
-    expected_non_res_pir_prediction_schema = StructType(
-        [
-            *non_res_pir_cleaned_ind_cqc_schema,
-            StructField(IndCQC.non_res_pir_linear_regression_model, FloatType(), True),
         ]
     )
 
@@ -4634,58 +4583,6 @@ class ValidateCleanedIndCqcData:
 
 
 @dataclass
-class ValidateImputedIndCqcAscwdsAndPir:
-    cleaned_ind_cqc_schema = StructType(
-        [
-            StructField(IndCQC.location_id, StringType(), True),
-            StructField(IndCQC.cqc_location_import_date, DateType(), True),
-        ]
-    )
-    imputed_ind_cqc_ascwds_and_pir_schema = StructType(
-        [
-            StructField(IndCQC.location_id, StringType(), True),
-            StructField(IndCQC.cqc_location_import_date, DateType(), True),
-            StructField(IndCQC.ascwds_workplace_import_date, DateType(), True),
-            StructField(IndCQC.cqc_pir_import_date, DateType(), True),
-            StructField(IndCQC.care_home, StringType(), True),
-            StructField(IndCQC.name, StringType(), True),
-            StructField(IndCQC.provider_id, StringType(), True),
-            StructField(IndCQC.provider_name, StringType(), True),
-            StructField(IndCQC.cqc_sector, StringType(), True),
-            StructField(IndCQC.registration_status, StringType(), True),
-            StructField(IndCQC.imputed_registration_date, DateType(), True),
-            StructField(IndCQC.dormancy, StringType(), True),
-            StructField(IndCQC.number_of_beds, IntegerType(), True),
-            StructField(
-                IndCQC.services_offered,
-                ArrayType(
-                    StringType(),
-                ),
-                True,
-            ),
-            StructField(IndCQC.primary_service_type, StringType(), True),
-            StructField(IndCQC.contemporary_ons_import_date, DateType(), True),
-            StructField(IndCQC.contemporary_cssr, StringType(), True),
-            StructField(IndCQC.contemporary_region, StringType(), True),
-            StructField(IndCQC.current_ons_import_date, DateType(), True),
-            StructField(IndCQC.current_cssr, StringType(), True),
-            StructField(IndCQC.current_region, StringType(), True),
-            StructField(IndCQC.current_rural_urban_indicator_2011, StringType(), True),
-            StructField(IndCQC.pir_people_directly_employed, IntegerType(), True),
-            StructField(IndCQC.establishment_id, StringType(), True),
-            StructField(IndCQC.organisation_id, StringType(), True),
-            StructField(IndCQC.total_staff_bounded, IntegerType(), True),
-            StructField(IndCQC.worker_records_bounded, IntegerType(), True),
-            StructField(IndCQC.ascwds_filled_posts_source, StringType(), True),
-            StructField(IndCQC.ascwds_filled_posts, DoubleType(), True),
-            StructField(IndCQC.ascwds_filled_posts_dedup_clean, DoubleType(), True),
-            StructField(IndCQC.pir_people_directly_employed_dedup, IntegerType(), True),
-        ]
-    )
-    calculate_expected_size_schema = cleaned_ind_cqc_schema
-
-
-@dataclass
 class ValidateCareHomeIndCqcFeaturesData:
     cleaned_ind_cqc_schema = StructType(
         [
@@ -4783,28 +4680,6 @@ class ValidateFeaturesNonResASCWDSWithoutDormancyIndCqcSchema:
         ]
     )
     non_res_ascwds_ind_cqc_features_schema = StructType(
-        [
-            StructField(IndCQC.location_id, StringType(), True),
-            StructField(IndCQC.cqc_location_import_date, DateType(), True),
-        ]
-    )
-
-    calculate_expected_size_schema = cleaned_ind_cqc_schema
-
-
-@dataclass
-class ValidateFeaturesNonResPirIndCqcSchema:
-    cleaned_ind_cqc_schema = StructType(
-        [
-            StructField(IndCQC.location_id, StringType(), True),
-            StructField(IndCQC.cqc_location_import_date, DateType(), True),
-            StructField(IndCQC.care_home, StringType(), True),
-            StructField(
-                IndCQC.imputed_non_res_pir_people_directly_employed, FloatType(), True
-            ),
-        ]
-    )
-    non_res_pir_ind_cqc_features_schema = StructType(
         [
             StructField(IndCQC.location_id, StringType(), True),
             StructField(IndCQC.cqc_location_import_date, DateType(), True),
@@ -4984,7 +4859,7 @@ class DiagnosticsOnKnownFilledPostsSchemas:
             StructField(IndCQC.non_res_with_dormancy_model, FloatType(), True),
             StructField(IndCQC.non_res_without_dormancy_model, FloatType(), True),
             StructField(IndCQC.non_res_combined_model, FloatType(), True),
-            StructField(IndCQC.non_res_pir_linear_regression_model, FloatType(), True),
+            StructField(IndCQC.imputed_pir_filled_posts_model, FloatType(), True),
             StructField(IndCQC.imputed_posts_non_res_combined_model, FloatType(), True),
             StructField(IndCQC.estimate_filled_posts, FloatType(), True),
             StructField(Keys.year, StringType(), True),
@@ -5010,7 +4885,7 @@ class DiagnosticsOnCapacityTrackerSchemas:
             StructField(IndCQC.non_res_with_dormancy_model, FloatType(), True),
             StructField(IndCQC.non_res_without_dormancy_model, FloatType(), True),
             StructField(IndCQC.non_res_combined_model, FloatType(), True),
-            StructField(IndCQC.non_res_pir_linear_regression_model, FloatType(), True),
+            StructField(IndCQC.imputed_pir_filled_posts_model, FloatType(), True),
             StructField(IndCQC.imputed_posts_non_res_combined_model, FloatType(), True),
             StructField(IndCQC.estimate_filled_posts, FloatType(), True),
             StructField(IndCQC.number_of_beds, IntegerType(), True),
