+++ utils/validation/validation_rules/estimated_ind_cqc_filled_posts_validation_rules.py
@@ -42,7 +42,7 @@ class EstimatedIndCqcFilledPostsValidationRules:
             IndCqcColumns.non_res_without_dormancy_model: 1.0,
             IndCqcColumns.number_of_beds: 1,
             IndCqcColumns.pir_people_directly_employed_dedup: 1,
-            IndCqcColumns.non_res_pir_linear_regression_model: 0.01,
+            IndCqcColumns.imputed_pir_filled_posts_model: 0.01,
             IndCqcColumns.posts_rolling_average_model: 1.0,
             IndCqcColumns.unix_time: 1262304000,  # 1st Jan 2010 in unix time
         },
@@ -57,7 +57,7 @@ class EstimatedIndCqcFilledPostsValidationRules:
             IndCqcColumns.non_res_without_dormancy_model: 3000.0,
             IndCqcColumns.number_of_beds: 500,
             # IndCqcColumns.pir_people_directly_employed_dedup: 3000, # temporarily removed until imputation is fixed
-            IndCqcColumns.non_res_pir_linear_regression_model: 3000.0,
+            IndCqcColumns.imputed_pir_filled_posts_model: 3000.0,
             IndCqcColumns.posts_rolling_average_model: 3000.0,
             IndCqcColumns.unix_time: int(time.time()),  # current unix time
         },
